resolve 34.36.20.125 as ping.remarkable.com
I did read your weg page before connecting my new remarkable to the
network for the first time, so I was warned about looking for this
specific DNS query. Any whatdontyouknow: here it is.

What I didn't do (yet?) is to try to see whether I can snoop at what is
going on: at the moment I do block all outgoing SYN requests, so the
connection is not even initiated. However, if the port (443) is an
indication for https and the remarkable checks certificates correctly,
that may not be that easy. This, the PR only "reverse-resolves" the IP
to its name.